novametrics-backend/requirements.txt
paho-mqtt>=2.0.0
python-dotenv
tzdata
flask
flask-cors
gunicorn